+++ strawberry/fastapi/constants.py
@@ -1,0 +1,13 @@
+NO_GRAPHIQL_MSG: str = "GraphiQL not enabled"
+GRAPHIQL_EXAMPLE: str = """
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Strawberry GraphiQL</title>
+    ...
+</head>
+<body>
+    ...
+</body>
+</html>
+"""
